Move BeginningOfLineAnchor into separate file
- This is consistent with all our other regexp mutators.
