+++ lib/mutant.rb
@@ -88,6 +88,7 @@ require 'mutant/mutator/node'
 require 'mutant/mutator/node/generic'
 require 'mutant/mutator/node/regexp'
 require 'mutant/mutator/node/regexp/alternation_meta'
+require 'mutant/mutator/node/regexp/beginning_of_line_anchor'
 require 'mutant/mutator/node/regexp/capture_group'
 require 'mutant/mutator/node/regexp/character_type'
 require 'mutant/mutator/node/regexp/end_of_line_anchor'
