+++ lib/mutant/mutator/node/regexp.rb
@@ -14,17 +14,6 @@ module Mutant
             children.each_index(&method(:mutate_child))
           end
         end # RootExpression
-
-        # Mutator for beginning of line anchor `^`
-        class BeginningOfLineAnchor < Node
-          handle(:regexp_bol_anchor)
-
-        private
-
-          def dispatch
-            emit(s(:regexp_bos_anchor))
-          end
-        end # BeginningOfLineAnchor
       end # Regexp
     end # Node
   end # Mutator
